requirements.txt
discord.py
python-dotenv
pymongo
bson